refactor(core): split embedding provider configs under 500 LOC
Move LocalConfig, Azure/Custom configs, and artifact verification out of
provider_config.rs so production sources stay within the LOC gate.
