+++ memory-core/src/embeddings/config/provider_config.rs
@@ -2,10 +2,11 @@
 
 use serde::{Deserialize, Serialize};
 
-use super::{OptimizationConfig, mistral::MistralConfig, openai::OpenAIConfig};
-
-#[allow(unused_imports)]
-use super::mistral::OutputDtype;
+use super::OptimizationConfig;
+use super::cloud_config::{AzureOpenAIConfig, CustomConfig};
+use super::local_config::LocalConfig;
+use super::mistral::MistralConfig;
+use super::openai::OpenAIConfig;
 
 /// Provider-specific configuration
 ///
@@ -29,259 +30,6 @@ pub enum ProviderConfig {
     /// Custom provider configuration
     #[serde(rename = "custom")]
     Custom(CustomConfig),
-}
-
-/// Configuration for local embedding providers
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LocalConfig {
-    /// Model name/path
-    pub model_name: String,
-    /// Embedding dimension
-    pub embedding_dimension: usize,
-    /// Optimization settings
-    #[serde(default)]
-    pub optimization: OptimizationConfig,
-    /// When true, allow mock embeddings for tests/dev (S1.5).
-    ///
-    /// Production defaults to `false`: if the real model cannot load, the
-    /// provider reports degraded-mock / unavailable rather than silently
-    /// advertising mock vectors as healthy (see `EmbeddingHealth`).
-    #[serde(default)]
-    pub allow_mock_fallback: bool,
-    /// Optional model revision / pin label (S1.5b / F4.3).
-    #[serde(default)]
-    pub model_revision: Option<String>,
-    /// Optional expected SHA-256 hex digest of the primary model artifact (S1.5b).
-    #[serde(default)]
-    pub expected_sha256: Option<String>,
-    /// Optional maximum allowed artifact size in bytes (S1.5b).
-    #[serde(default)]
-    pub max_artifact_bytes: Option<u64>,
-}
-
-impl LocalConfig {
-    /// Create a new local config.
-    ///
-    /// Mock fallback defaults to **enabled** so offline tests/dev keep working.
-    /// Call [`Self::with_allow_mock_fallback`]`(false)` for production fail-closed
-    /// loading (S1.5). Mock mode never reports production-ready via `is_available`.
-    #[must_use]
-    pub fn new(model_name: impl Into<String>, dimension: usize) -> Self {
-        Self {
-            model_name: model_name.into(),
-            embedding_dimension: dimension,
-            optimization: OptimizationConfig::local(),
-            allow_mock_fallback: true,
-            model_revision: None,
-            expected_sha256: None,
-            max_artifact_bytes: None,
-        }
-    }
-
-    /// Enable mock embeddings for tests and local development (S1.5).
-    #[must_use]
-    pub fn with_allow_mock_fallback(mut self, allow: bool) -> Self {
-        self.allow_mock_fallback = allow;
-        self
-    }
-
-    /// Pin model revision string (S1.5b).
-    #[must_use]
-    pub fn with_model_revision(mut self, revision: impl Into<String>) -> Self {
-        self.model_revision = Some(revision.into());
-        self
-    }
-
-    /// Require SHA-256 hex digest of the model file (S1.5b).
-    #[must_use]
-    pub fn with_expected_sha256(mut self, digest_hex: impl Into<String>) -> Self {
-        self.expected_sha256 = Some(digest_hex.into());
-        self
-    }
-
-    /// Reject artifacts larger than `max_bytes` (S1.5b).
-    #[must_use]
-    pub fn with_max_artifact_bytes(mut self, max_bytes: u64) -> Self {
-        self.max_artifact_bytes = Some(max_bytes);
-        self
-    }
-
-    /// Get the effective embedding dimension
-    #[must_use]
-    pub fn effective_dimension(&self) -> usize {
-        self.embedding_dimension
-    }
-}
-
-/// Verify a model artifact against optional digest and size limits (S1.5b / F4.3).
-///
-/// # Errors
-/// Returns an error when the file is missing, oversized, or digest mismatches.
-pub fn verify_model_artifact(
-    path: &std::path::Path,
-    expected_sha256: Option<&str>,
-    max_bytes: Option<u64>,
-) -> anyhow::Result<()> {
-    use sha2::{Digest, Sha256};
-    use std::fs;
-    use std::io::Read;
-
-    let meta = fs::metadata(path).map_err(|e| {
-        anyhow::anyhow!("model artifact metadata failed for {}: {e}", path.display())
-    })?;
-    let len = meta.len();
-    if let Some(max) = max_bytes {
-        if len > max {
-            anyhow::bail!(
-                "model artifact {} is {len} bytes, exceeds max_artifact_bytes={max}",
-                path.display()
-            );
-        }
-    }
-    if let Some(expected) = expected_sha256 {
-        let expected = expected.trim().to_ascii_lowercase();
-        let mut file = fs::File::open(path)
-            .map_err(|e| anyhow::anyhow!("open model artifact {}: {e}", path.display()))?;
-        let mut hasher = Sha256::new();
-        let mut buf = [0_u8; 8192];
-        loop {
-            let n = file
-                .read(&mut buf)
-                .map_err(|e| anyhow::anyhow!("read model artifact {}: {e}", path.display()))?;
-            if n == 0 {
-                break;
-            }
-            hasher.update(&buf[..n]);
-        }
-        let actual = format!("{:x}", hasher.finalize());
-        if actual != expected {
-            anyhow::bail!(
-                "model artifact {} digest mismatch: expected {expected}, got {actual}",
-                path.display()
-            );
-        }
-    }
-    Ok(())
-}
-
-impl Default for LocalConfig {
-    fn default() -> Self {
-        // Offline/dev default: mock allowed. Production should set
-        // `allow_mock_fallback: false` and treat only `EmbeddingHealth::Real`
-        // as production-ready.
-        Self::new("sentence-transformers/all-MiniLM-L6-v2", 384)
-    }
-}
-
-/// Configuration for Azure OpenAI Service
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AzureOpenAIConfig {
-    /// Deployment name
-    pub deployment_name: String,
-    /// Resource name
-    pub resource_name: String,
-    /// API version
-    pub api_version: String,
-    /// Embedding dimension
-    pub embedding_dimension: usize,
-    /// Optimization settings
-    #[serde(default)]
-    pub optimization: OptimizationConfig,
-}
-
-impl AzureOpenAIConfig {
-    /// Create a new Azure OpenAI config
-    #[must_use]
-    pub fn new(
-        deployment_name: impl Into<String>,
-        resource_name: impl Into<String>,
-        api_version: impl Into<String>,
-        dimension: usize,
-    ) -> Self {
-        Self {
-            deployment_name: deployment_name.into(),
-            resource_name: resource_name.into(),
-            api_version: api_version.into(),
-            embedding_dimension: dimension,
-            optimization: OptimizationConfig::azure(),
-        }
-    }
-
-    /// Get the Azure OpenAI endpoint URL
-    #[must_use]
-    pub fn endpoint_url(&self) -> String {
-        format!(
-            "https://{}.openai.azure.com/openai/deployments/{}/embeddings?api-version={}",
-            self.resource_name, self.deployment_name, self.api_version
-        )
-    }
-
-    /// Get the effective embedding dimension
-    #[must_use]
-    pub fn effective_dimension(&self) -> usize {
-        self.embedding_dimension
-    }
-}
-
-/// Configuration for custom embedding providers
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CustomConfig {
-    /// Model identifier
-    pub model_name: String,
-    /// Embedding dimension
-    pub embedding_dimension: usize,
-    /// Base URL for the API
-    pub base_url: String,
-    /// Custom endpoint path (optional)
-    #[serde(default)]
-    pub api_endpoint: Option<String>,
-    /// Optimization settings
-    #[serde(default)]
-    pub optimization: OptimizationConfig,
-}
-
-impl CustomConfig {
-    /// Create a new custom config
-    #[must_use]
-    pub fn new(
-        model_name: impl Into<String>,
-        dimension: usize,
-        base_url: impl Into<String>,
-    ) -> Self {
-        Self {
-            model_name: model_name.into(),
-            embedding_dimension: dimension,
-            base_url: base_url.into(),
-            api_endpoint: None,
-            optimization: OptimizationConfig::local(),
-        }
-    }
-
-    /// Set custom API endpoint
-    #[must_use]
-    pub fn with_endpoint(mut self, endpoint: impl Into<String>) -> Self {
-        self.api_endpoint = Some(endpoint.into());
-        self
-    }
-
-    /// Get the full embeddings endpoint URL
-    #[must_use]
-    pub fn embeddings_url(&self) -> String {
-        let endpoint = self.api_endpoint.as_deref().unwrap_or("/embeddings");
-        let base = self.base_url.trim_end_matches('/');
-
-        if endpoint.starts_with('/') {
-            format!("{base}{endpoint}")
-        } else {
-            format!("{base}/{endpoint}")
-        }
-    }
-
-    /// Get the effective embedding dimension
-    #[must_use]
-    pub fn effective_dimension(&self) -> usize {
-        self.embedding_dimension
-    }
 }
 
 impl ProviderConfig {
@@ -368,16 +116,7 @@ impl ProviderConfig {
     pub fn local_default() -> Self {
         Self::Local(LocalConfig::default())
     }
-}
 
-impl Default for ProviderConfig {
-    fn default() -> Self {
-        Self::local_default()
-    }
-}
-
-// Convenience constructors for common configurations
-impl ProviderConfig {
     /// OpenAI text-embedding-3-small
     #[must_use]
     pub fn openai_3_small() -> Self {
@@ -421,8 +160,15 @@ impl ProviderConfig {
     }
 }
 
+impl Default for ProviderConfig {
+    fn default() -> Self {
+        Self::local_default()
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use super::super::mistral::OutputDtype;
     use super::*;
 
     #[test]
@@ -453,30 +199,10 @@ mod tests {
     }
 
     #[test]
-    fn test_azure_openai_endpoint() {
-        let config = AzureOpenAIConfig::new("my-deployment", "my-resource", "2023-05-15", 1536);
-        assert_eq!(
-            config.endpoint_url(),
-            "https://my-resource.openai.azure.com/openai/deployments/my-deployment/embeddings?api-version=2023-05-15"
-        );
-    }
-
-    #[test]
-    fn test_custom_config_url() {
-        let config = CustomConfig::new("custom-model", 768, "https://api.example.com/v1")
-            .with_endpoint("/custom-embeddings");
-        assert_eq!(
-            config.embeddings_url(),
-            "https://api.example.com/v1/custom-embeddings"
-        );
-    }
-
-    #[test]
     fn test_provider_config_serialization() {
         let config = ProviderConfig::openai_3_small();
         let json = serde_json::to_string(&config).unwrap();
 
-        // Verify the tag is present - use either quoted or unquoted format
         assert!(
             json.contains("\"provider\":\"openai\"") || json.contains("\"openai\""),
             "Expected provider tag in JSON, got: {json}"
